GeomOpt

Documentation for GeomOpt.

GeomOpt.cartesian_to_fractional

GeomOpt.clamp_atoms

GeomOpt.construct_optimization_function

GeomOpt.fractional_to_cartesian

GeomOpt.get_optimizable_mask

GeomOpt.get_positions

GeomOpt.get_positions_cart

GeomOpt.mask_vector_list

GeomOpt.n_optimizable_coordinates

GeomOpt.set_optimizable_mask

GeomOpt.update_optimizable_coordinates

GeomOpt.update_optimizable_coordinates_cart

GeomOpt.update_positions

GeomOpt.update_positions_cart

GeomOpt.cartesian_to_fractional — Method

cartesian_to_fractional(system::AbstractSystem, positions::AbstractVector{<:AbstractVector{<:Real}})

Given a list of cartesian coordinates, convert them to fractional.

GeomOpt.clamp_atoms — Method

Clamp given atoms if the system. Clamped atoms are fixed and their positions will not be optimized. The atoms to be clamped should be given as a list of indies corresponding to their positions in system.particles.

GeomOpt.construct_optimization_function — Method

By default we work in cartesian coordinaes.
Note that internally, when optimizing the cartesian positions, atomic units 
are used.

GeomOpt.fractional_to_cartesian — Method

fractional_to_cartesian(system::AbstractSystem, positions::AbstractVector{<:AbstractVector{<:Real}})

Given a list of fractional coordinates, convert them to cartesian.

GeomOpt.get_optimizable_mask — Method

get_optimizable_mask(system::AbstractSystem) -> AbstractVector{<:AbstractVector{<:Bool}}

Returns the optimizable mask of the system (see documentation for set_optimizable_mask.

GeomOpt.get_positions — Method

get_positions(system::AbstractSystem, positions::Vector{<:AbstractVector{<:Real}})

Returns positions, in fractional coordinates, of the system's particles.

# Returns
- `Vector{Vector{L}}`

GeomOpt.get_positions_cart — Method

get_positions_cart(system::AbstractSystem, positions::Vector{<:AbstractVector{<:Real}})

Returns positions, in cartesian coordinates, of the system's particles.

# Returns
- `Vector{Vector{L}}`

GeomOpt.mask_vector_list — Method

Given a list of vectors and a mask, return the masked version in a one-dimensional list. This is used to concatenate the forces returned in the scfres into a single list, that can be used by the optimizer.

GeomOpt.n_optimizable_coordinates — Method

n_optimizable_coordinates(system::AbstractSystem)

Returns the number of optimizable coordinates in the system.

GeomOpt.set_optimizable_mask — Method

set_optimizable_mask(system::AbstractSystem, mask::AbstractVector{<:AbstractVector{<:Bool}})

Sets the masks defining which coordinates of the system can be optimized. The mask is a list of 3-d boolean vectors, where, for each particle of the system, the vector specifies which of the xyz coordinates can be optimized by flagging them with true.

GeomOpt.update_optimizable_coordinates — Method

update_optimizable_coordinates(system::AbstractSystem, positions::Vector{<:AbstractVector{L}}) where {L <: Real}

Creates a new system based on $system$ with the optimizable coordinates are updated to the ones provided (in the order in which they appear in system.particles).

GeomOpt.update_optimizable_coordinates_cart — Method

update_optimizablecoordinates_cart(system::AbstractSystem, positions::Vector{<:AbstractVector{<:L}}) where {L <: Unitful.Length}

Creates a new system based on $system$ with the optimizable coordinates are updated to the ones provided (in the order in which they appear in system.particles), cartesian coordinates version..

GeomOpt.update_positions — Method

update_postions(system::AbstractSystem, positions::Vector{<:AbstractVector{<:Real}})

Creates a new system based on $system$ but with atoms positions updated to the ones specified in positions. Note that we work in fractional coordinates.

GeomOpt.update_positions_cart — Method

update_postions_cart(system::AbstractSystem, positions::Vector{<:AbstractVector{<:Real}}) where {L <: Unitful.Length}

Creates a new system based on $system$ but with atoms positions updated to the ones specified in positions, using cartesian coordinates.