GeomOpt

Documentation for GeomOpt.

GeomOpt.DofManager

GeomOpt.analyze_mask

GeomOpt.DofManager — Type

DofManager:

Constructor:

DofManager(sys; variablecell = false, r0 =..., free=..., clamp=..., mask=...)

variablecell determines whether the cell is fixed or allowed to change during optimization

r0 is a reference length-scale, default set to one (in the unit of sys), this is used to non-dimensionalize the degrees of freedom.

In addition set at most one of the kwargs:

no kwarg: all atoms are free

free : list of free atom indices (not dof indices)

clamp : list of clamped atom indices (not dof indices)

mask : 3 x N Bool array to specify individual coordinates to be clamped

Meaning of dofs

On call to the constructor, DofManager stores positions and cell X0, C0, dofs are understood relative to this initial configuration. get_dofs(sys, dm::DofManager) returns a vector that represents the non-dimensional displacement and a deformation matrix (U, F). The new configuration extracted from a dof vector is understood as

The new cell: C = F * C0

The new positions: 𝐫[i] = F * (X0[i] + U[i] * r0)

One aspect of this definition is that clamped atom positions still change via the deformation F. This is natural in the context of optimizing the cell shape.

GeomOpt.analyze_mask — Method

analyze_mask : helper function to generate list of dof indices from lists of atom indices indicating free and clamped atoms